requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.0.74
pillow==10.0.0
streamlit==1.25.0
streamlit-drawable-canvas==0.9.3